Added implementation file for Fruit class
