+++ CMakeLists.txt
@@ -25,5 +25,6 @@ target_sources(
   Sources/Inky.cpp
   Sources/Pinky.cpp
   Sources/Clyde.cpp
+  Sources/Fruit.cpp
 )
 target_link_libraries(pacman_sdl ${SDL2_LIBRARIES} SDL2::Main SDL2::Image SDL2::TTF SDL2::Mixer)
